fix(analytics): replace flows with afterCsr and remove helpers
