+++ core/packages/analytics/pwa/client.js
@@ -1,7 +1,6 @@
 import components from './components';
 import env from './env';
-import flow from './flows/client';
 import Store from '../shared/stores';
 
 export default components;
-export { Store, flow, env };
+export { Store, env };
