+++ core/packages/analytics/shared/stores/index.js
@@ -1,7 +1,8 @@
-import { types, getRoot } from 'mobx-state-tree';
+import { types, getRoot, addMiddleware } from 'mobx-state-tree';
 import GoogleAnalytics from './google-analytics';
 import GoogleTagManager from './google-tag-manager';
 import ComScore from './comscore';
+import { afterAction } from './utils';
 
 const Analytics = types
   .model('Analytics')
@@ -26,6 +27,15 @@ const Analytics = types
     sendEvent(event) {
       self.googleAnalytics.sendEvent(event);
       self.googleTagManager.sendEvent(event);
+    },
+    afterCsr() {
+      const { connection } = getRoot(self);
+      // Send pageviews when route has changed
+      const pageViewMiddleware = afterAction(
+        'routeChangeSucceed',
+        self.sendPageView,
+      );
+      addMiddleware(connection, pageViewMiddleware);
     },
   }));
