+++ core/packages/analytics/shared/stores/utils.js
@@ -1,5 +1,11 @@
-/* eslint-disable import/prefer-default-export */
 import { getRoot } from 'mobx-state-tree';
+import sha256 from 'crypto-js/sha256';
+import base64 from 'crypto-js/enc-base64';
+
+export const afterAction = (name, callback) => (call, next) => {
+  next(call);
+  if (call.type === 'action' && call.name === name) callback(call);
+};
 
 export const generateEvent = self => event => {
   const { connection } = getRoot(self);
@@ -16,4 +22,13 @@ export const generateEvent = self => event => {
     action,
     label,
   };
+};
+
+export const getRoute = ({ page }) =>
+  typeof page === 'number' ? 'list' : 'single';
+
+export const getHash = (site, selectedItem) => {
+  const { type, id, page } = selectedItem;
+  const data = JSON.stringify([site, type, id, page]);
+  return base64.stringify(sha256(JSON.stringify(data))).slice(0, 19);
 };
